+++ sdk/communication/communication-identity/samples/v1/javascript/issueToken.js
@@ -22,17 +22,16 @@ async function main() {
 
   // Create user
   console.log("Creating User");
-
   const user = await client.createUser();
-
   console.log(`Created user with id: ${user.communicationUserId}`);
   console.log("Issuing Token");
 
   // Issue token and get token from response
-  const vanillaToken = await client.getToken(user, scopes);
-  console.log(`Issued token: ${vanillaToken.token}`);
+  const defaultToken = await client.getToken(user, scopes);
+  console.log(`Issued token: ${defaultToken.token}`);
 
   // Issue token with custom expiration and get token from response
+  console.log("Issuing Token with custom expiration.");
   const tokenOptions = { tokenExpiresInMinutes: 60 };
   const { token, expiresOn } = await client.getToken(user, scopes, tokenOptions);
   console.log(`Issued token with custom expiration: ${token}`);