+++ sdk/communication/communication-identity/CHANGELOG.md
@@ -7,8 +7,8 @@
     - `createUserAndToken` and `getToken` methods now accept `CreateUserAndTokenOptions` and `GetTokenOptions` that provide the ability to create a Communication Identity access token with custom expiration via optional `tokenExpiresInMinutes` property:
         - `createUserAndToken(scopes: TokenScope[], options: CreateUserAndTokenOptions = {})`
         - `getToken(user: CommunicationUserIdentifier, scopes: TokenScope[], options: GetTokenOptions = {})`
-        - `export declare interface CreateUserAndTokenOptions extends OperationOptions { tokenExpiresInMinutes?: number; }`
-        - `export declare interface GetTokenOptions extends OperationOptions { tokenExpiresInMinutes?: number; }`
+        - `interface CreateUserAndTokenOptions extends OperationOptions { tokenExpiresInMinutes?: number; }`
+        - `interface GetTokenOptions extends OperationOptions { tokenExpiresInMinutes?: number; }`
     - Added a new API version `2022-10-01` that is now the default API version.
 
 ## 1.1.1 (Unreleased)