+++ sdk/communication/communication-identity/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 1.2.0 (Unreleased)
+## 1.2.0 (2022-10-12)
 
 ### Features Added